// frontend/src/api/generated/model/userCreateRequest.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * Your API
 * OpenAPI spec version: 1.0.0
 */
import type { UserCreateRequestAttributes } from './userCreateRequestAttributes';

export interface UserCreateRequest {
  name?: string;
  email?: string;
  password?: string;
  phoneNumber?: string;
  roleIds?: number[];
  groupIds?: number[];
  attributes?: UserCreateRequestAttributes;
}